TiMa/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
paho-mqtt==1.6.1
python-dotenv==1.0.1
APScheduler==3.10.4
pytest==8.1.1
